Add support for navigating between pages
This uncovered a wrong assumption about how Instant works in regard to subscriptions. When the data is already there, then the result is returned synchronously in the callback, which means that the query variable must be set prior to calling the subscribe function. (which is used to check if the query is still the active one.)